services/telemetry-service/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.36.1
pymongo==4.15.0
pydantic==2.12.3
pydantic-settings==2.11.0
python-dateutil==2.9.0.post0
